// src/api/__generated__/story/model/pageDtoUserProfileUnifiedResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { UserProfileUnifiedResponse } from './userProfileUnifiedResponse';

/**
 * 游标分页结果
 */
export interface PageDtoUserProfileUnifiedResponse {
  /** 本页请求的分页大小 */
  pageSize?: number;
  /** 下一页游标；hasMore=false 时为 -1 */
  mark?: string;
  /** 数据列表 */
  list?: UserProfileUnifiedResponse[];
  /** 是否还有下一页 */
  hasMore?: boolean;
}