// src/api/__generated__/story/model/userProfileUnifiedResponseType.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

/**
 * Display type discriminator: DRAMA / DRAMA_EPISODE / SHORT_VIDEO
 */
export type UserProfileUnifiedResponseType =
  (typeof UserProfileUnifiedResponseType)[keyof typeof UserProfileUnifiedResponseType];

export const UserProfileUnifiedResponseType = {
  DRAMA: 'DRAMA',
  DRAMA_EPISODE: 'DRAMA_EPISODE',
  SHORT_VIDEO: 'SHORT_VIDEO',
} as const;