// src/api/__generated__/wallet/model/userProfileResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { WalletAddressResponse } from './walletAddressResponse';

/**
 * 用户资料
 */
export interface UserProfileResponse {
  /** 用户ID */
  userId?: string;
  /** 登录类型：1-Privy托管登录 2-Web3钱包登录 */
  loginType?: string;
  /** 主钱包地址 */
  walletAddress?: string;
  /** 用户钱包地址列表 */
  walletAddresses?: WalletAddressResponse[];
  /** 邮箱 */
  email?: string;
  /** 用户昵称 */
  nickname?: string;
  /** 用户头像URL */
  avatarUrl?: string;
  /** 用户自己的邀请码 */
  inviteCode?: string;
  /** 邀请人用户ID */
  inviterUserId?: string;
  /** 是否跳过邀请码：0-否 1-是 */
  skipInviteCode?: string;
  /** 用户状态：1-正常 0-禁用 */
  status?: string;
  /** 用户trust */
  trust?: string;
  /** 用户简介 */
  profile?: string;
  /** 是否删除：0-未删除 1-已删除（仅标记，不影响任何逻辑） */
  isDeleted?: string;
}